Refactor: Move setup out of saga
